OCP-26014: fix syntax and fix clean-up steps (#2681)
There seems to be a syntax issue and a test logic issue.

I don't know why/how sometimes the old step "OVN is functional on the cluster"
works, but the proper syntax seems to be to require a "Given" before the step.

When the step fails without the "Given", it fails in steps_parser.

```
<NoMethodError: undefined method `each' for nil:NilClass>
/home/jenkins/.gem/bundler/ruby/2.7.0/gems/cucumber-5.3.0/lib/cucumber/runtime/support_code.rb:21:in `steps'
/home/jenkins/.gem/bundler/ruby/2.7.0/gems/cucumber-5.3.0/lib/cucumber/gherkin/steps_parser.rb:23:in `parse'
/home/jenkins/.gem/bundler/ruby/2.7.0/gems/cucumber-5.3.0/lib/cucumber/runtime/support_code.rb:63:in `invoke_dynamic_steps'
/opt/rh/rh-ruby27/root/usr/share/ruby/forwardable.rb:235:in `invoke_dynamic_steps'
/home/jenkins/.gem/bundler/ruby/2.7.0/gems/cucumber-5.3.0/lib/cucumber/glue/proto_world.rb:168:in `block (2 levels) in for'
/home/jenkins/.gem/bundler/ruby/2.7.0/gems/cucumber-5.3.0/lib/cucumber/glue/proto_world.rb:41:in `steps'
features/step_definitions/meta_steps.rb:48:in `block (2 levels) in '
lib/base_helper.rb:160:in `wait_for'
features/step_definitions/meta_steps.rb:43:in `/^I wait(?: up to ([0-9]+|<%=.+?%>) seconds)? for the steps to pass:$/'
features/networking/pod.feature:496:in `I wait up to 60 seconds for the steps to pass:'
```

But the real issue seems to be the test logic.  We don't delete the pod to force CNO
to repair the cluster until after the scenario so the
"OVN is functional on the cluster" step won't pass.

Instead move the delete pod and OVN functional check into the clean-up
steps.

CNO recovers the pods quickly in less than 60 seconds